refactor: single event guest and user split
