+++ client/routes.js
@@ -7,7 +7,7 @@ import {Signup} from './components/auth-form-su'
 import {me} from './store'
 import HomePage from './Pages/Homepage'
 import AllEvents from './components/AllEvents'
-import SingleEvent from './components/SingleEvent'
+import SingleEvent from './components/SingleEvent/SingleEvent'
 import Profile from './Pages/Profile'
 import AddEvent from './Pages/AddEvent'
 import EventsByCategory from './components/EventByCategory'
