+++ client/store/rsvp.js
@@ -2,6 +2,7 @@ import axios from 'axios'
 
 const LOAD_ATTENDEES = 'LOAD_ATTENDEES'
 const CHECK_IN = 'CHECK_IN'
+const IS_RSVPED = 'IS_RSVPED'
 
 const checkIn = checkedIn => ({
   type: CHECK_IN,
@@ -9,9 +10,12 @@ const checkIn = checkedIn => ({
 })
 const loadAttendees = attendees => ({type: LOAD_ATTENDEES, attendees})
 
+const isRSVPed = rsvp => ({type: IS_RSVPED, rsvp})
+
 const initialState = {
   attendees: [],
-  checkedIn: false
+  checkedIn: false,
+  rsvped: false
 }
 
 export const findCheckIn = eventId => async dispatch => {
@@ -33,12 +37,23 @@ export const getAttendees = eventId => async dispatch => {
   }
 }
 
+export const findRsvp = eventId => async dispatch => {
+  try {
+    const {data} = await axios.get(`/api/rsvp/userrsvp/${eventId}`)
+    dispatch(isRSVPed(data))
+  } catch (error) {
+    console.log(error)
+  }
+}
+
 const rsvpReducer = (state = initialState, action) => {
   switch (action.type) {
     case LOAD_ATTENDEES:
       return {...state, attendees: action.attendees}
     case CHECK_IN:
       return {...state, checkedIn: action.checkedIn}
+    case IS_RSVPED:
+      return {...state, rsvped: action.rsvp}
     default:
       return state
   }
